requirements.txt
streamlit
langchain
openai
requests
pandas
yfinance 
python-dotenv 